% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_obj_gen_funcion.R
\name{coef_obj_gen}
\alias{coef_obj_gen}
\title{Title Obtencion del vector de coeficientes del problema.}
\usage{
coef_obj_gen(lp)
}
\arguments{
\item{lp}{Un objeto de programacion lineal generado conel paquete lpSolveAPI}
}
\value{
Devuelve un vector de tantos elementos como columnas tenga el problema
}
\description{
Title Obtencion del vector de coeficientes del problema.
}
\examples{
print("Ya meteremos algo!")
}
