% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_celdas_funcion.R
\name{coord_celdas}
\alias{coord_celdas}
\title{Obtencion de coordenadas de celdas en una hoja de cálculo}
\usage{
coord_celdas(wb, regionName)
}
\arguments{
\item{wb}{Donde se coloca un objeto Workbook de XLConnect}

\item{regionName}{Nombre de la region del archivo .xlsx de la que se desea extraer las coordenadas}
}
\value{
Devuelve las coordenadas de todas las celdas que componen la region solicitada en el archivo .xlsx.
}
\description{
Utiliza las regiones creadas dentro del workbook con las funciones writeNamedRegion del paquete XLConnect
}
\examples{
print("Introduciremos ejemplos posteriormente, disculpa las molestias.")
}
