% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaCeldasScript.R
\name{formula_celdas}
\alias{formula_celdas}
\title{Generacion de formulas de Simplex dentro del archivo .xlsx}
\usage{
formula_celdas(mat_celdas, pivot_list)
}
\arguments{
\item{mat_celdas}{Matriz con las coordenadas de la region donde se desean calcular las formulas.}

\item{pivot_list}{Objeto que contiene las coordenadas del pivote, tanto numericas como textuales.}
}
\value{
Devuelve la formulas pertinentes para la region, siguiendo la regla del cuadrado y teniendo en cuenta aquella celda donde se encuentre el pivote.
}
\description{
Generacion de formulas de Simplex dentro del archivo .xlsx
}
\examples{
print("Introduciremos ejemplos posteriormente, disculpa las molestias.")
}
