% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_cb_function.R
\name{index_cb_calc}
\alias{index_cb_calc}
\title{Calculate index of cb coefficients}
\usage{
index_cb_calc(lp, matriz_restr_start)
}
\arguments{
\item{lp}{An object of class \code{lpExtPtr} from the lpSolveAPI package}

\item{matriz_restr_start}{The initial restriction matriz from the Simplex problem}
}
\value{
A numeric vector for indexing purposes
}
\description{
Returns the index for the basis coefficients inside a Simplex problem
}
\details{
This function identifies all those coefficients which are a part of the
initial basis from the specified linear problem
}
