% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_prob_function.R
\name{invert_prob}
\alias{invert_prob}
\title{Transformador de problemas primales a duales}
\usage{
invert_prob(lp)
}
\arguments{
\item{lp}{Un objeto lpPtrExtr completo, resuelto o no.}
}
\value{
La version dual del objeto introducido.
}
\description{
Transformador de problemas primales a duales
}
\examples{
library(lpSolveAPI)
problem_80 <- make.lp(3,4)

set.objfn(problem_80, c(-8, -1, 0, 7))
set.rhs(problem_80, c(2, 1, 8))
set.constr.type(problem_80, c("=", ">=", "<="))
set.row(problem_80, 1, c(2, 5, -4, 1))
set.row(problem_80, 2, c(7, 0, 6, -7))
set.row(problem_80, 3, c(-7, -8, 3, -6))

solve(problem_80)
problem_80
invert_prob(problem_80)
}
