% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz_restr_function.R
\name{matriz_restr_calc}
\alias{matriz_restr_calc}
\title{Extract the initial restriction matrix from a lpExtPtr object}
\usage{
matriz_restr_calc(lp)
}
\arguments{
\item{lp}{An object of class \code{lpExtPtr} from the lpSolveAPI package}
}
\value{

}
\description{
Takes a simplex problem defined by lpSolveAPI and then computes the initial
restriction matrix
}
