% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_calc_funcion.R
\name{pivot_calc}
\alias{pivot_calc}
\title{Calculo del pivote de una iteracion de Simplex}
\usage{
pivot_calc(mat_restr, mat_restr_celdas, cost_reducidos, mat_M, mat_rhs, cual_M,
  workbook, lp)
}
\arguments{
\item{mat_restr}{Matriz numerica de restricciones del problema Simplex a resolver}

\item{mat_restr_celdas}{Matriz de coordenadas de la matriz de restricciones del archivo .xlsx}

\item{cost_reducidos}{Vector de los costes reducidos generados en una iteracion de Simplex durante el criterio de entrada}
}
\value{
Una lista con tres elementos: Fila, Columna y Coordenada
}
\description{
Calculo del pivote de una iteracion de Simplex
}
\examples{
print("Introduciremos ejemplos posteriormente, disculpa las molestias.")
}
