% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablaSimplexfuncion.R
\name{tablaSimplex}
\alias{tablaSimplex}
\title{Title Generador de soluciones de problemas de programacion lineal con simplex en tablas de formato .xlsx}
\usage{
tablaSimplex(lp, wb)
}
\arguments{
\item{lp}{Objeto tipo lpExtPtr proveniente del paquete lpSolveAPI y que debe estar resuelto con la funcion solve antes de ser introducido}

\item{wb}{Donde se coloca un objeto Workbook de XLConnect}
}
\value{
Rellena un archivo .xlsx, previamente creado con loadWorkbook, con la resolucion de un problema de Simplex.
}
\description{
Title Generador de soluciones de problemas de programacion lineal con simplex en tablas de formato .xlsx
}
\examples{
print("Introduciremos ejemplos posteriormente, disculpa las molestias.")
}
