% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marco_tabla_simplex_funcion.R
\name{texteitor}
\alias{texteitor}
\title{Title Generacion de marcos con texto y regiones para el problema lineal planteado}
\usage{
texteitor(lp, wb, iteracion)
}
\arguments{
\item{lp}{Objeto tipo lpExtPtr proveniente del paquete lpSolveAPI y que debe estar resuelto con la funcion solve antes de ser introducido}

\item{wb}{Donde se coloca un objeto Workbook de XLConnect}

\item{iteracion}{Numero de la tabla que se va a generar dentro del archivo .xlsx}
}
\value{
Establece las regiones de cada una de las partes del problema de programacion lineal y los marcos indicativos textuales.
}
\description{
Title Generacion de marcos con texto y regiones para el problema lineal planteado
}
\examples{
print("Introduciremos ejemplos posteriormente, disculpa las molestias.")
}
